Refactor methods from gluon/admin to gluon/fileutils
